+++ LICENSE
@@ -1,1 +1,27 @@
-TO DO
+Copyright (c) 2023 Guillermo Gomez
+
+# Ana Paula (Licenciada en Derecho Ana Paula) License
+
+This software is the property of its authors and is provided as an Artificial Intelligence (AI) service that functions as a professional paralegal specializing in various areas of Mexican law.
+
+## Usage
+
+Users are authorized to use this software to aid in understanding legal documents and Mexican law concepts. This includes, but is not limited to ranch owners, college students, and experienced lawyers. The software can communicate in both Spanish and English, as well as other languages to accommodate expats or immigrants living in Mexico.
+
+## Limitations
+
+The software is not a licensed attorney and, as such, it is explicitly prohibited for the software to provide legal advice. This tool aims to simplify and explain complex legal language and concepts; it does not replace professional legal consultation.
+
+## Data Protection
+
+This software adheres to the Mexican law 'LEY FEDERAL DE PROTECCIÓN DE DATOS PERSONALES EN POSESIÓN DE LOS PARTICULARES', ensuring all data shared in the conversation is protected. It does not keep any record of the data, and every conversation is destroyed when the conversation is closed.
+
+## Disclaimer
+
+The authors of this software are not responsible for any misunderstandings, errors, or legal issues that may arise from the use of this software. The software is continually learning and improving, and its capabilities are subject to change.
+
+## Acknowledgment
+
+By using this software, the user acknowledges the software's limitations as an AI and agrees not to use the software to provide legal advice. The user also recognizes the software's commitment to data protection under Mexican law.
+
+For further information, please contact the authors of this software.
