--- datasets/'Cake hoop': la tendencia más cool en decoración de pasteles de boda.txt
@@ -1,43 +1,0 @@
-Todo lo que necesitas saber sobre el banquete de boda
-'Cake hoop': la tendencia más cool en decoración de pasteles de boda 
-¿Están buscando una decoración original para el pastel de boda y sienten que ya lo han visto todo? Conozcan los 'cake hoops', una base ornamental para pasteles en varios niveles que dará un giro de 360° a la manera de presentar su tarta nupcial.
-
-Verónica Ledesma            
-
-
-
-El pastel que elijan para su boda puede llegar a ser inolvidable para sus invitados a causa de su sabor o de su maravillosa presentación. Los  son una base decorativa que puede adornarse con variaciones infinitas de flores, listones y colores para ajustarse al concepto de su celebración y llenar de encanto su tarta nupcial. ¿Listos para descubrir esta tendencia en decoración de pasteles de boda? ¡Seguro les encanta!
-¿Qué es un ?
-Materiales de los 
-Qué formas y colores pueden tener los 
-¿Cómo se puede decorar el ?
-¿En qué tipo de boda pueden usarse?
-Un , también conocido como aro para pastel, es una estructura que sirve para decorar y presentar una tarta. Básicamente, consta de una base de la cual nacen uno o dos aros que sirven para enmarcar y decorar el diseño. En resumen, un  es un marco para el pastel de boda, que puede decorarse de diferentes maneras. 
-marco para el pastel de boda
-
-
-Los materiales más comunes de los  son la madera, el metal o una combinación de los dos. Estas bases pueden encontrarse en tiendas de decoración o ser adquiridas directamente con su repostero, aunque también su proveedor de arreglos florales podrá ayudarlos.
-a madera, el metal o una combinación de los dos
-proveedor de arreglos florales
-La plataforma sobre la que se coloca el pastel puede ser visible o no. Las bases metálicas suelen ser delgadas y se disimulan fácilmente, pero también existen gruesas bases de madera clara u oscura que lucen elegantes en las mesas de postres. 
-Por otro lado, existen aros de distintas alturas, todo depende de los niveles y del tamaño que tendrá su pastel de boda. Los aros del pastel pueden medir desde 40 centímetros y alcanzar hasta un metro de altura.
- existen aros de distintas alturas
-
-
-La mayoría de los  tienen forma circular y pueden ser blancos, rosas, dorados, plateados o negros. Para crear armonía con esa forma circular, sobre la base se colocan pasteles con esa misma figura. Elijan el tono del aro que se ajuste mejor a la paleta de colores que utilizarán en su boda.
-blancos, rosas, dorados, plateados o negros
-Existen bases con dos aros, que pueden ser paralelos, estar cruzados o tomar la figura de un arco. Hay otros que adquieren la forma heptagonal o piramidal. Sus ángulos dan un aspecto moderno a la mesa de postres, solo asegúrense de que estas formas geométricas sean armónicas con el resto de la decoración de su boda. También hay bases que constan de niveles separados para colocar varios pasteles en forma de torre. Cada nivel del pastel está separado por aros en sentido vertical para crear la apariencia de pasteles flotantes.
-bases con dos aros
-forma heptagonal o piramidal
-
-
-Además de darle una vista encantadora al pastel de boda, los  pueden decorarse con follaje y con flores. Las guías de vegetación pueden enroscarse a lo largo de la circunferencia totalmente, o cargarse hacia uno de los lados. Los detalles florales podrán dar acentos localizados o colocarse en especie de botones en zonas específicas.
-decorarse con follaje y con flores
-En los aros pequeños se recomienda concentrar la decoración en la parte inferior y aligerarla conforme se recorre hacia arriba, de este modo la atención se irá hacia el pastel y se creará un efecto visual de mayor altura. Los aros más grandes pueden ser ornamentados por completo con guías vegetales, listones de seda o encaje, o bien con discretas series de luces. Incluso puede colgárseles una tela vaporosa y transparente. Si el pastel es mucho menor en proporción respecto al aro, pueden adornar la parte superior al colgar velas de tipo pecera. El brillo del cristal y la luz de la flama es encantador.
-guías vegetales, listones de seda o encaje, o bien con discretas series de luces
-Elijan las flores y el follaje de acuerdo con lo aquello que presentarán en los arcos florales, en los centros de mesa y en la mesa de novios. La estructura y el pastel pueden colocarse en una mesa solitaria o en la mesa de postres. Las velas, linternas y más flores servirán para completar la escena.
-
-
-El uso de rosas, encaje o tela para decorar un será ideal en bodas románticas, en las que también abunden las velas, las transparencias y las luces cálidas. Las bodas glamorosas serán discretas y sutiles en los ornamentos, apenas detallados.
-Una boda en la playa alternaría grandes hojas de monstera, palma o filodendro con flores tropicales por la extensión del arco, que en las celebraciones más modernas tomará la figura heptagonal. Por último, el de las bodas vintage ostentará colores pálidos, como el amarillo o el rosa deslavado, en la decoración. Se adornará con flores preservadas o con hojas de la Pampa, además de que reposará sobre una mesa de madera envejecida.
-Más de un invitado se detendrá en este  a admirar la decoración de su pastel de boda. Una vez que haya pasado el baile de novios, las foto con la tarta no podrán faltar.
