--- datasets/'Mocktails': 12 cocteles sin alcohol para bodas... pero con mucho sabor.txt
@@ -1,54 +1,0 @@
-Todo lo que necesitas saber sobre el banquete de boda
-'Mocktails': 12 cocteles sin alcohol para bodas... pero con mucho sabor
-Impresionen a sus invitados y mímenlos con estas propuestas de 'mocktails' que caerán muy bien en su boda. La combinación de fruta, helado, leches y cremas da como resultado bebidas sin alcohol no solo atractivas a la vista, también para el paladar.
-
-Gabriela Acosta            
-
-
-
-Ya sea por estilo de vida, por culto religioso o por gusto hay personas que no consumen bebidas con alcohol. Incluir en la barra de boda  (cocteles sin alcohol) o servirlos junto a la mesa de dulces será un gran detalle y lucirán muy divertidos. Además, si tienen invitados pequeños, hasta ellos podrán brindar con estas divertidas fórmulas. Si quieren saber con qué pueden preceder a la comida o ayudar a digerirla, estos ricos cocteles sin alcohol no pueden faltar en su boda.
-Dicen que la primera impresión no se olvida. Y por algo duraron semanas seleccionando todos los detalles que mejor se verían en su recepción, soñaron con manteles de una docena de colores y analizaron con lupa el menú que después servirán. Por esta misma razón, reciban a sus invitados con unas deliciosas y frescas bebidas que anticipen el buen gusto que imperará durante el resto de la celebración. Por su composición, estos son los cocteles sugeridos para la recepción. 
-bebidas que anticipen el buen gusto que imperará
-Este licuado de durazno, jugo de naranja y jugo de mango volverá loco de gusto a quien lo pruebe. Como su nombre lo dice, el Sun summer es ideal para el día y en bodas en verano, ya que las frutas con que se elabora son de temporada. Si quieren una presentación de impacto, pidan que se sirva con una bola de helado de mandarina, ¡delicioso!
-para el día y en bodas en verano
-
-
-Si se decantaron por ofrecer una piña colada en su preparación original, inclúyanla en la versión sin alcohol. Los presentes estarán gustosos de probar esta combinación de leche de coco, piña y crema de leche. Su gran versatilidad le permite estar en barras de día, de noche, interiores o exteriores. Además, queda bien en cualquier temporada. 
-gran versatilidad
-Entre los cocteles sin alcohol fáciles se encuentra este que, muy probable, sea la bebida de gusto general, porque es dulce y con cierto sabor a cítricos, pero no tanto como para empalagar. En su versión sin vino y sin cualquier otro alcohol, la sangría se prepara con agua mineral, azúcar, limón, naranja, jugo de uva y manzanas cortadas en pequeños trozos. Seguramente su proveedor de catering para boda les ofrecerá increíbles propuestas de presentación. ¡No olviden que se toma bien fría!
-de gusto general, porque es dulce y con cierto sabor a cítricos
-catering para boda
-
-
-Si están buscando un coctel con frambuesa y quieren prescindir del tequila y el ron, este es el ideal. Nadie, ni ustedes, podrá resistirse a este coctel sin alcohol para boda que es el resultado de licuar leche de coco con fresas, frambuesas, duraznos y piñas. Su tonalidad naranja llamará la atención a distancia; un Acapulco sin alcohol seguro conseguirá que sus invitados hagan fila.
-leche de coco con fresas, frambuesas, duraznos y piñas
-tonalidad naranja
-Es cierto que una michelada sin cerveza no es una michelada, por eso pueden prepararla con cerveza sin alcohol, limón, salsa inglesa, sal, pimienta, jugo de tomate y salsa Tabasco. Esta bebida se sugiere para bodas en la playa o de estilo campestre y es ideal para refrescar tanto de día como de noche. Podrían acompañar con camarón o cecina a modo de botana. ¡Riquísimo!
-para bodas en la playa o de estilo campestre
-Este coctel de frutos rojos sin licor será la estrella de su celebración nupcial. La combinación de sabores dará como resultado una bebida atractiva, deliciosa y rica en antioxidantes. ¡Perfecto para que ningún invitado se quede quieto en su silla!
-La receta consiste en mezclar 50 gramos de zarzamoras, frambuesas y arándanos con un poco de jarabe de azúcar. Para decorar, agreguen coco y arándanos al gusto.
-zarzamoras, frambuesas y arándanos
-
-
-Este coctel es muy amigable porque no contiene fruta de temporada, ya que el plátano lo encontrarán en cualquier época. Acompañado de crema de coco, leche evaporada y hielo, su textura y sabor conquistarán a más de un paladar. Si quieren que el sabor sea más intenso, añadan jugo de piña. Para decorar, una rodaja de lima y una linda flor que armonice con su decoración nupcial serán suficiente. 
-en cualquier época
-Esta versión de mojito apta para todos prescinde del ron y se sirve con jugo de limón, hojas de menta, almíbar, agua mineral, hierbabuena y hielo. No hay duda de que todos sus ingredientes refrescan y les dejarán un rico sabor de boca.
-todos sus ingredientes refrescan
-
-
-La fiesta ya está en su mayor auge y sus invitados necesitan hidratarse para tomar energía. Ofrezcan opciones frescas, pero igualmente deliciosas y que sirvan de digestivo para que continúen la celebración con el mismo ímpetu con el que comenzaron.
-deliciosas y que sirvan de digestivo
-Los cítricos son buenos para la hidratación y la digestión, este es un coctel compuesto por jugo de arándano, piña, gotas de limón y agua mineral. Si los sabores les atraen y quieren tener también su versión con alcohol, deben saber que un Flamingo normalmente lleva ron blanco.
-Así como se arriesgaron con su original look, ahora denle otro toque juguetón a sus cocteles con una exótica, pero sabrosa propuesta. El Indian style es todo un contraste de sabores: mitad fresco, mitad picante. Incluye jengibre, agua tónica y pimienta. ¿Se atreverían a incluirlo?
-exótica, pero sabrosa propuesta
-
-
-
-
-Entre las recetas más buscadas de cocteles sin alcohol para boda en la playa está sex on the beach. No importa si el sol ya se ocultó, igual la disfrutarán los invitados. La receta original tiene vodka, pero en su versión sin alcohol consta de jugo de durazno, pera, piña, hojas de menta y una rodaja de limón para el toque final.
-durazno, pera, piña, hojas de menta y una rodaja de limón
-Después de gastar energía, a nadie le caería mal un  de té verde con jugo de naranja, toronja y limón. Para quienes cuidan su figura, este será el coctel deseado.
- de té verde con jugo de naranja, toronja y limón
-
-
-Aquí está la prueba de que los cocteles no solo son ricos si llevan alcohol, aunque eso sí, deberán olvidarse de contar las calorías con estas bebidas. Si ustedes son de los más preocupados por este tema, para contrarrestar pueden buscar entre los proveedores alguno que ofrezca opciones bajas en azúcar. ¿Les convence la sugerencia?
