--- datasets/'Wedding brunch': la tendencia más cool para parejas modernas.txt
@@ -1,60 +1,0 @@
-Todo lo que necesitas saber sobre el banquete de boda
-'Wedding brunch': la tendencia más cool para parejas modernas 
-Los novios actuales apuestan por celebraciones menos tradicionales y se dejan conquistar por nuevas tendencias. Por ejemplo, reemplazar el banquete clásico por uno más informal como el 'brunch'. ¿Lo conocen? Les contamos sus principales ventajas.
-
-Gabriela Acosta            
-
-
-
-Las diferentes alternativas que hay para celebrar una boda permiten cumplir las necesidades y gustos de cada pareja. Así, hoy en día, los novios más prácticos sustituyen los pasteles de boda por  y algunas prometidas lucen vestidos de novia cortos desafiando el protocolo más tradicional. Lo mismo sucede con la gastronomía. El  ha sido tendencia entre quienes quieren casarse en un ambiente íntimo y algo desenfadado. ¿Quieren saber de qué trata?
-¿Qué es un ?
-¿En qué tipo de celebraciones va bien un ?
-
-La mejor opción para la tornaboda
-
-
- viene de la combinación de las palabras  (desayuno) y (almuerzo o comida), por lo que en términos prácticos significa que es un banquete entre el desayuno y la comida. En cuanto a la cantidad de alimentos, es más fuerte que el desayuno, pero más ligero que la comida.
-banquete entre el desayuno y la comida
-Su origen tiene varias teorías. Por un lado, se cree que se deriva de cuando las madres británicas preparaban desayunos tardíos para sus hijos que trabajaban de noche. Otros consideran que el concepto nace debido a que los deportistas de alto rendimiento necesitaban un segundo desayuno, más potente, para recuperar energía tras un duro entrenamiento. Y, por último, hay quien apunta a que fue en Nueva York del siglo XIX, cuando surgió como remedio para la resaca dominical.
-origen tiene varias teorías
-madres británicas preparaban desayunos tardíos
-deportistas de alto rendimiento necesitaban un segundo desayuno
-Nueva York del siglo XIX
-remedio para la resaca dominical
-Sea como sea, este concepto se ha vuelto tendencia en eventos y bodas en las que, poco a poco, las clásicas comilonas han ido aligerándose un poco. Esto se debe a que el  es un tipo de banquete práctico, moderno, relajado y más económico. 
-las clásicas comilonas han ido aligerándose
-práctico, moderno, relajado y más económico
-
-
-Debido a su naturaleza, el  es ideal para celebraciones íntimas. Se sugiere que inicie entre las 11 de la mañana y la 1 de la tarde. Consideren comunicar este aspecto en sus invitaciones de boda, ya que al ser un número reducido de personas, la puntualidad será muy importante. Sobre todo, porque implicará que la ceremonia será más temprano.
-ideal para celebraciones íntimas
-inicie entre las 11 de la mañana y la 1 de la tarde
-Al no ser una celebración formal el código de vestimenta es mucho más relajado. Sus invitados lucirán perfectos con vestidos para fiesta de día y trajes no tan solemnes.
-código de vestimenta es mucho más relajado
-En cuanto a la locación, con un  resulta mucho más fácil encontrar fechas disponibles y lugares en donde festejar su matrimonio. Puede ser desde una terraza o un jardín hasta un lindo restaurante para bodas. ¿Creen que puedan caber en uno que sea significativo para ustedes, como el elegido para la entrega del anillo de compromiso?
-más fácil encontrar fechas disponibles y lugares
-restaurante para bodas
-Otra ventaja del  de boda es que no será tan necesario que cada invitado tenga un asiento asignado. Al ser menos formal y siempre que el menú lo permita, también es propicio que las personas convivan de pie, como en un coctel de tipo buffet. Pese a esto, sí hay que considerar intercalar entre las estaciones de comida algunas salas  y otros espacios con mobiliario clásico que garanticen que, al menos, el 70% pueda estar sentado para comer. Además de acatar las medidas de seguridad conforme al aforo y a la distancia entre las personas que indican los protocolos sanitarios. 
-no será tan necesario que cada invitado tenga un asiento asignado
-como en un coctel de tipo buffet
-
-
-La finalidad de una  es tener una reunión más personal, pero no por eso se deben dejar de lado detalles como la decoración, distribución y, por supuesto, el sabor. Por eso es importante considerar dentro del menú del  una variedad de platillos salados y dulces, así como bebidas frías y calientes para antes o después de la comida. Consulten con diferentes proveedores de catering para boda que les podrán dar las mejores ideas para combinar todos los sabores que se les ocurran para su boda. 
-variedad de platillos salados y dulces
-proveedores de catering para boda
-Los platos salados pueden ser ensaladas, tostadas, bocadillos, empanadas, hamburguesas o canapés. También, huevos en sus distintas cocciones y combinaciones acompañados de papas asadas o fritas, así como otras guarniciones. Para acompañar, un amplio abanico de aderezos, salsas, frutos secos y distintos tipos de pan serán necesarios. 
-ensaladas, tostadas, bocadillos, empanadas, hamburguesas o canapés
-huevos en sus distintas cocciones 
-distintos tipos de pan 
-La leche y sus derivados no pueden faltar. Una tabla de quesos, yogures, así como mantequillas y cremas deben estar dentro del menú.
-leche y sus derivados
-En los platos dulces, consideren una variedad de frutas y o crepas con mermeladas y jarabes que serán complementos perfectos. Para el postre, obviamente no puede faltar el pastel de boda, así como cuernos, galletas, ,  y repostería en general como helados, dulces y confituras.
-frutas y o crepas 
-Entre las bebidas, incluyan jugos, cocteles frutales, aguas frescas, batidos y malteadas, infusiones, cafés, refrescos, cervezas y vino. Aunque el consumo de vino será significativamente menor que en una fiesta clásica nocturna, inclúyanlo en el menú de bebidas y para el brindis. También un espumoso será bien recibido y refrescante para las horas centrales del día.
-jugos, cocteles frutales, aguas frescas, batidos y malteadas, infusiones, cafés, refrescos, cervezas y vino
-un espumoso será bien recibido
-
-
-Si no quieren perderse la oportunidad de celebrar en grande su enlace nupcial, entonces dejen el  de boda para el día siguiente. Es una excelente opción para convivir con sus seres más allegados y una forma de consentirlos con una deliciosa tornaboda. Eso sí, asegúrense de notificar a quienes están contemplados para esta continuación del festejo cuando entreguen las invitaciones de boda elegantes. ¡No querrán perdérselo!
-dejen el  de boda para el día siguiente
-¿Les gustó la idea del ? Quizá ya habían escuchado el término o incluso disfrutado de alguno en su restaurante favorito. Si optan por sumarse a esta tendencia y celebrar una  también ustedes podrán lucir un look nupcial más relajado. Si su anhelo es adoptar nuevas tradiciones, adáptenlas al lugar y momento. ¡Echen a volar su imaginación! Por ejemplo, lanzar el ramo al salir de su ceremonia civil o religiosa, ¡será toda una sorpresa!
