--- datasets/10 anillos de compromiso para 10 novias: ¿cuál es para ella?.txt
@@ -1,75 +1,0 @@
-La novia
-10 anillos de compromiso para 10 novias: ¿cuál es para ella?
-Todas las joyas son hermosas y más cuando sellan el amor entre dos personas. Para complacer a tu pareja con un anillo especial que vaya acorde a su personalidad, no dejes de leer la siguiente lista. ¡Será muy fácil encontrar el ideal para ella!  
-
-Gabriela Acosta            
-
-
-
-Recibir el anillo de compromiso ya es de por sí emocionante. Recibir uno que demuestre el tiempo que dedicó esa persona para encontrar la pieza especial para ella lo hace aún más valioso. Si es alocada, no sigue las reglas y te ha hecho saber que cuando llegue el día llevará un vestido de novia corto, la joya para comprometerse no será la tradicional. Pero ¿cómo saber cuál es el anillo indicado? La música que escucha, los peinados de novia que le llamaron la atención y su carácter son buenas pistas para empezar a buscar y estas descripciones también. ¡Suerte! 
-anillo de compromiso
-vestido de novia corto
-peinados de novia
-
-
-No hay joya más romántica que aquella que está elaborada en oro rosa. Si a tu futura prometida le gusta seguir las tendencias, una sortija que combine este metal con los diamantes en pavé será ideal para ella. ¿Te estás preguntando cuánto cuesta un anillo de compromiso como este? Debido a que el color rosa se obtiene de una mezcla de 75% de oro puro, 20% de cobre y 5% de plata, resulta ser más económico que los que son de oro puro. Lo que elevará el precio será la cantidad y pureza de las piedras que lleve el anillo en este tipo de mosaico, que se ha vuelto muy popular en los últimos años.
-le gusta seguir las tendencias
-cuánto cuesta un anillo de compromiso
-cantidad y pureza de las piedras
-
-
-Si aún no descubres en qué mano va el anillo de compromiso, puede que este anillo te haga dudar aún más de la respuesta. Los anillos para dos dedos están hechos para novias que se salen de lo común. ¿Qué mejor manera de simbolizar un compromiso fuerte como el que tienen? Por cierto, se vale cuestionar al joyero en qué dedos se llevan estas sortijas; ¿dónde crees que lo llevará tu pareja?
-en qué mano va el anillo de compromiso
-anillos para dos dedos
-
-
-Algunas mujeres tienen la superstición de que llevar perlas el día de la boda traerá mala suerte al matrimonio. Pero, ¿qué crees? En algunas culturas, esta piedra orgánica está relacionada directamente con el amor, la belleza y la fertilidad, por lo que, si ella es atrevida y nada supersticiosa, este anillo será el indicado. Seguro que sabrá complementar su look con unos aretes tipo broquel en esta piedra y que luzcan con su peinado recogido. Las propuestas con platino y oro blanco son muy lindas y elegantes. 
-el amor, la belleza y la fertilidad
-aretes tipo broquel
-peinado recogido
-con platino y oro blanco
-
-
-La alegría también se manifiesta en forma de lágrimas y este diseño le recordará siempre la felicidad explosiva que sintió cuando sellaron su compromiso con esta joya. Además, quedará perfecto con su personalidad creativa e independiente. Si su gusto es discreto y elegante, elige un anillo de compromiso de oro blanco o cualquier otro metal en este color, como la plata o el platino. Si es más arriesgada y vive bajo sus propias reglas, busca uno cuya piedra sea una esmeralda o un zafiro.
-personalidad creativa e independiente
-anillo de compromiso de oro blanco
-
-
-¡Hasta que la muerte los separe! Este pensamiento de amor para tu novia es uno de tantos que estarán presentes en su boda. Hay quienes graban palabras de amor en el interior de la banda y quienes deciden expresarlo con un anillo que, como este, lleve en el centro el símbolo del infinito. Si tu chica es sencilla, uno con aro delgado en donde destaque la figura de la eternidad la dejará encantada. Si es moderna y le gusta romper esquemas, opta por uno que, en lugar de un aro, tenga una serie de símbolos infinitos.
-pensamiento de amor para tu novia
-símbolo del infinito
-
-
-¿Conoces la tradición que dicta que la novia debe llevar algo viejo, algo prestado, algo azul y algo nuevo? Ahórrale la búsqueda de los dos últimos elementos con un anillo con piedra azul, como el zafiro. Si entregarás el anillo en la playa, elige una aguamarina para entonar con el paisaje. ¡Se impactará ante tanta belleza! Además, este color tiene un lindo significado, pues se cree que el azul ayuda a la buena comunicación en el matrimonio y simboliza la pureza, la fidelidad y la confianza.
-algo viejo, algo prestado, algo azul y algo nuevo
-zafiro
-aguamarina
-la pureza, la fidelidad y la confianza
-
-
-Es verdad que el aro completo simboliza la eternidad, pero por qué no pensar que las alianzas abiertas permiten la entrada del amor por siempre. Además de que su diseño permite ajustarse mejor a la anchura del dedo (siempre con mucha delicadeza), este tipo de diseños conquista a las chicas poco convencionales, que no se conforman con lo establecido y que disfrutan las nuevas propuestas. Demuéstrale que dedicaste tiempo para encontrar el anillo ideal para ella.
-permiten la entrada del amor por siempre
-ajustarse mejor a la anchura del dedo 
-Aquellos que desembocan en diamantes son espectaculares, pero también los encontrarás con muchos diseños en los extremos. Estrellas, corazones, tréboles, flechas y notas musicales son solo algunos ejemplos de lo que la joyería de autor puede ofrecerles.
-joyería de autor
-
-
-Seguramente en más de alguna ocasión han platicado sobre cómo les gustaría casarse. Si ella te ha compartido que anhela un vestido de novia corte princesa y una boda de ensueño, no dudes ni por un segundo que sus ojos se llenarán de alegría cuando le entregues el clásico anillo de compromiso. No pasa de moda y se mantiene en el gusto de las novias más románticas.
-vestido de novia corte princesa
-No pasa de moda
-las novias más románticas
-
-
-Ya sea por gusto o por presupuesto, no todos los anillos de compromiso tienen que llevar necesariamente diamantes. Hay piedras preciosas y semipreciosas que, además de bellas, son fuertes y resistentes para ser incrustadas en una joya como esta. Los zafiros, rubíes y esmeraldas son las más comunes, pero hay otras, como la turquesa, la amatista, el granate, el topacio, el zircón, la aguamarina, la morganita y el ópalo que te ayudarán a expresar todo tu amor. Además, te permitirán descubrir anillos de compromiso de todos los precios. Que tu pareja tenga alergia al oro también puede ser suficiente motivo para buscar una hermosa alianza en plata. 
-zafiros, rubíes y esmeraldas
-la turquesa, la amatista, el granate, el topacio, el zircón, la aguamarina, la morganita y el ópalo
-anillos de compromiso de todos los precios
-
-
-No podrás esperar a ver la cara de tu novia cuando le propongas matrimonio con un anillo de compromiso Tiffany como este. ¡Diamantes por aquí, por allá y por todos lados! Si tu presupuesto te lo permite y quieres sorprenderla con un anillo que desde lejos anuncie su compromiso, ¡los diamantes nunca fallan!
-anillo de compromiso Tiffany
-Si tu presupuesto te lo permite
-¿Lograste identificar el anillo ideal para tu chica? Si necesitas consultar más modelos, no te quedes sin ver en la web el catálogo de joyería nupcial. Quedará fascinada con la elección que hagas y querrá presumir con pensamientos de amor y a la menor oportunidad su compromiso. Ahora solo queda planear una entrega del anillo de compromiso de impacto: ¿será algo clásico, una sorpresa o consideras que será mejor hablar tranquilamente de su futuro en pareja? Con la fórmula que escojas, disfruten juntos de todas estas emociones de planear la boda ¡y descubran todas las herramientas que Bodas.com.mx les brinda en este camino!
-catálogo de joyería nupcial
-pensamientos de amor
-entrega del anillo de compromiso
