--- datasets/10 claves para llegar descansados al gran día: ¡la boda se acerca!.txt
@@ -1,35 +1,0 @@
-Tratamientos
-10 claves para llegar descansados al gran día: ¡la boda se acerca!
-Durante mucho tiempo trabajarán duro para tener la boda de sus sueños, por lo que se merecen tener el ánimo y la energía para disfrutarla, así que no dejen de lado las siguientes alternativas para que lleguen como nuevos al altar.
-
-Érika Flores            
-
-
-
-El ajetreo por los preparativos y resolver lo necesario para tener todo tiempo los traerá un poco de cabeza, así que necesitarán hacer un alto para recuperar energía y sacudirse el estrés. Hay muchas formas de lograrlo, por eso les enlistamos algunas a fin de que disfruten sin nervios o estrés antes de la boda.
- 
-Merecen un buen descanso. Su ritmo de vida cambió después de anunciar su compromiso, pues los preparativos de la boda les demandaron más tiempo y esfuerzo día a día. Al menos dos semanas antes de la celebración asegúrense de que su descanso sea reparador: tomen siestas de 15 minutos por la tarde, eviten acostarse después de las 11:00 de la noche para que duerman al menos siete horas y a la hora de ir a la cama eviten ver la televisión, la computadora, el celular o cualquier aparato que emita luz porque les frustrará el sueño.
-eviten acostarse después de las 11:00 de la noche
-
-
-Si están a tiempo de planear mejor la organización de la boda, lleven una agenda diaria en la que establezcan metas que puedan cumplir para que la satisfacción de haberlo logrado los haga dormir tranquilos, a la par que van eliminando pendientes y no tengan que correr en la semana previa a la boda.
- establezcan metas
-Si falta poco para el gran día y aún les falta mucho por resolver, no duden en pedir unos días en su trabajo para que se dediquen de lleno y puedan encargarse de varias cosas en una sola jornada, a fin de que les quede al menos una semana sin pendientes y puedan descansar mejor.
-Sabemos que desean que todo salga perfecto y sentirán la necesidad de ocuparse de todo, pero habrá muchas tareas que fácilmente podrán delegar, como la solicitud de cotizaciones, la elaboración de centros de mesa, la organización de los invitados en la recepción y algunas compras de productos o servicios que ustedes ya hayan decidido.
-
-
-Sí, suena muy trillado, pero es verdad: alimentarse sanamente será básico para su salud, no solo en las semanas previas a la boda, pues estarán bajo mucho estrés y si se saltan las comidas o le entran duro a las harinas, el azúcar y las grasas — sobre todo por la noche — les afectará. Lo mejor será tomar suficiente agua natural para mantenerse cuerdos e hidratados, e incluir a su dieta alimentos que los carguen de energía como el huevo, plátano, avena, espinacas y los frutos secos, especialmente entre comidas, además de verduras y carnes magras.
-tomar suficiente agua natural
-Como si se tratara de magia, al descargar el estrés se sentirán ligeros y renovados; una excelente forma de lograrlo es el ejercicio. Nadar, correr y andar en bicicleta son de las rutinas más efectivas, aunque todo es bienvenido, pues toda actividad física relaja al cuerpo y disminuye la ansiedad. ¡Vayan a box! Cada que golpeen el costal o la pera se sentirán liberados. Ojo, no se excedan: deben llegar sanos y salvos al día más importante de sus vidas.
-
-
-Escuchar música, ir al cine, jugar juegos de mesa con amigos, leer, armar rompecabezas, tejer, bailar, cantar... Todos tenemos una actividad favorita, así que, por favor, no la dejen de lado.  Eso sí, eviten a toda costa las desveladas y el alcohol.
-Quizá no sea un hábito común, pero les caerán de perlas unas sesiones de yoga, pilates, meditación o ejercicios de respiración, actividades maravillosamente efectivas para recuperar la paz y el descanso.
-
-
-Reír los hará muy felices; además de trabajar los músculos del abdomen, la risa oxigenará mejor su organismo, fortalecerá su sistema de defensas, controlará la ansiedad  y apartará los pensamientos depresivos. Qué mejor si se sientan juntos a ver programas o películas de comedia que los tenga a carcajadas al menos una vez al día.
-Disfruten del fresco aroma que emiten los árboles y las flores en el área verde más cercana a casa, mientras cantan los pajarillos. Caminar con los pies descalzos sobre el pasto les generará una agradable sensación de bienestar. Procuren hacerlo por la mañana en un sitio donde haya poco ruido y puedan concentrarse en sus pensamientos más positivos. 
-
-
-Si está dentro de sus posibilidades, contraten a un profesional de belleza o aparten fecha y hora para un tratamiento que incluya masajes o baños de vapor. Valdrá la pena visitar un spa al menos dos veces en las semanas previas a la boda para que, además de masaje, les apliquen tratamientos que mejoren la salud y el aspecto de su piel. Otra opción es el temazcal, donde podrán deshacerse de los tóxicos que les haya generado el estrés.
-¿Qué les parecieron estas alternativas? ¿Ya saben con cuál empezar? La mejor estrategia para evitar que lleguen agotados a a la boda es que organicen sus nupcias tan pronto como puedan, después de haber celebrado la entrega del anillo de compromiso porque el tiempo es oro. No descarten la contratación de un wedding planner, pues si desean participar más activamente en la planeación, pueden establecer las tareas del organizador con algo que se ajuste a su presupuesto.
