--- datasets/Costo promedio de las cosas de boda.txt
@@ -1,24 +1,0 @@
-This is the cost of all items that we have for a wedding in average
-Salon de fiesta por 7 horas : $100000
-DJ : $25500
-Iglesia : $5000
-Tequila para 100 personas : $23022
-Cerveza : $2500
-Vino Tinto : $5449
-Anillos de Boda : $6000
-Maquillaje de la novia : $5000
-Fotográfo : $21600
-Ramos : $3000
-Coro : $3800
-Smoking : $4050
-Decoración : $9750
-Mariachi  : $9600
-Flores : $11000
-Comida de Desvelados : $1393
-Invitaciónes Digitales : $2436
-Cilindros : $839
-Zapatos del novio : $3000
-Zapatos de la novia : $4000
-Banquete bistró a 3 tiempos para 100 personas : $61000
-Mobiliario : $7000
-Valet Parking por 7 horas : $5900
