--- datasets/La etiqueta o código de vestimenta para invitados hombres a una boda.txt
@@ -1,61 +1,0 @@
-Los invitados de la boda
-La etiqueta o código de vestimenta para invitados hombres a una boda
-Fuiste invitado a una boda y te han señalado un código de vestimenta. ¿Cómo puedes elegir el tipo de traje adecuado, el color, los zapatos y los accesorios idóneos? Te presentamos algunas pautas e ideas para elegir el look de invitado perfecto. 
-
-Verónica Ledesma            
-
-
-
-Tienes una cita para un evento especial pero aún no has decido cómo vestir. Los novios pueden haberte lanzado una clave al indicar el código de vestimenta para bodas en su invitación, pero también hay parejas que esperan que lo deduzcas o que desean que sus invitados se sientan cómodos y vistan de la manera más auténtica. ¡La solución es fácil! Sabrás qué usar una vez que consideres tres cosas en relación con el evento: el lugar, la hora y el tipo de ceremonia que se realizará. De manera que, si te preocupa saber cómo debe ir vestido a una boda un hombre, te explicamos todo a continuación.
-De gala: el máximo nivel de etiqueta
-Padrinos o testigos en una boda de etiqueta rigurosa
-¿Y el invitado común de etiqueta rigurosa?
-Etiquetas de vestimenta para hombres en boda formal
-Código de vestimenta para boda tipo coctel
-Etiqueta de vestir para boda en la playa
-
-
-El más alto grado de etiqueta se expresa en el código de gala, que implica el uso del frac. En lenguaje coloquial, se conoce el frac como el “traje de pingüino”.
-implica el uso del frac
-El frac es corto y sigue una línea horizontal en la parte del frente, con botonadura simple o cruzada que se lleva abierta. En la parte trasera, los faldones tienen una forma apuntada. Se viste con pajarita de piqué blanco y chaleco del mismo color. La camisa blanca se cierra en los puños con gemelos, y asoma fuera de la chaqueta. Los zapatos deben ser de charol con cordones, y con un pequeño tacón.
-En países de Europa como España, este código se utiliza en cenas de Estado o bailes de gala, que generalmente se celebran de noche en interiores. En ese contexto, no es usual llevar frac en una celebración de bodas. En cambio, en Latinoamérica hay ocasiones en las que los contrayentes masculinos visten frac, y el resto de los invitados lleva esmoquin, a excepción de padrinos o testigos. Como invitado, considera el frac solo si los novios te lo solicitan, siempre serán ellos quienes marquen el protocolo a seguir.
-considera el frac solo si los novios te lo solicitan
-
-En una boda de etiqueta rigurosa (o formal rigurosa) los contrayentes masculinos usarán chaqué, y es posible que soliciten vestir una versión adaptada de su indumentaria a sus padrinos o testigos, que son invitados especiales. El chaqué puede usarse de día o de noche. La levita es corta en el frente y tiene un faldón que se divide en dos piezas redondeadas en la parte trasera que llega a la altura de las rodillas.
-El chaqué puede usarse de día o de noche
-Por otro lado, la semilevita puede entenderse como una versión más actual de la levita clásica. Es más corta y ajustada de manera adecuada al cuerpo, genera una silueta atractiva que distingue a quienes lo portan.
-El chaleco que se viste con levita o semilevita, puede usarse con y sin solapas. Si lleva solapas, estas son redondeadas, el chaleco se cruza en el frente, es recto en la parte inferior y lleva dos hileras de seis botones. En cambio, el chaleco sin solapas tiene una sola hilera de botones y termina en forma triangular en la parte inferior. Utiliza las líneas y cortes de esta prenda para estilizar tu figura.
-El chaleco, la corbata de seda o pajarita pueden ser del color deseado. Puede usarse un pantalón liso del color de la levita negra en su versión más clásica, aunque también puede ser una levita gris o azul; también puede usarse con un pantalón de rayas discretas. Los zapatos deben ser sin brillo y se usan modelos con cordones.
-Puede usarse un pantalón liso del color de la levita negra
-Encuentra el traje perfecto
-Encuentra el traje perfecto
-
-
-En una boda en la que el código de vestimenta para bodas es formal riguroso los invitados pueden considerar el esmoquin, que, a diferencia del traje normal, lleva solapas de seda. La camisa blanca de cuello de paloma oculta los botones, puede usarse con moño o corbatín con un fajín del mismo color. Otro accesorio que no puede faltar son las mancuernillas.
-pueden considerar el esmoquin
-El pantalón siempre es negro y puede llevar un galón lateral de seda que se coordina con las solapas del mismo tejido que lleva la chaqueta. Se viste con zapatos de charol o de cordones. La chaqueta puede ser de otros colores, aunque en la versión más usual, es negra.
-El pantalón siempre es negro
-
-
-En una boda formal la corbata es imprescindible. El traje puede constar de dos o tres piezas, es decir, puede llevar chaleco, o no. Puedes combinar un saco y pantalón a juego con un chaleco de color arena, gris, celeste, rosa, burdeos y hasta verde. 
-El traje puede constar de dos o tres piezas
-Si el evento es de día, elige un traje de tonos claros como el gris. Reserva los colores oscuros para eventos nocturnos, como el azul tinta o el gris Oxford. El negro clásico de noche siempre viene excelente.
-Reserva los colores oscuros para eventos nocturnos
-Además de los tonos, puedes jugar con la textura de los tejidos para adaptar el conjunto a tu estilo. Los patrones a cuadros en el saco, o en saco y pantalón, son otra opción además del tartán, el patrón Príncipe de Gales, pata de gallo o .
-No te olvides de accesorios como el pañuelo o el fistol. Deberás elegir uno ellos para no excederte en los ornamentos. La variedad de combinaciones es infinita.
-
-
-El código semiformal de día o casual, puede prescindir de la corbata. El traje debe ser de tonos claros y puedes integrar tejidos ligeros como el lino o jugar con el color de tu camisa. El coctel de noche implica el uso de colores oscuros, pero ya que se trata de una ocasión semiformal, deberás evitar el negro. Un pañuelo liso o estampado en el bolsillo te dará un toque único.
-puede prescindir de la corbata
- jugar con el color de tu camisa
-
-
-Una vez más, el estilo de la boda de playa lo dictarán los novios. En esas ocasiones, la guayabera o la camisa de lino en tonos crudos pueden ser la opción, no olvides que la guayabera se usa desfajada. También puedes vestir una camisa de algodón y pantalones de lino en colores claros.
-la guayabera o la camisa de lino
-en tonos crudos
-Olvida las sandalias, ya que se trata de un evento especial. Los mocasines te permitirán proyectar una actitud relajada sin perder el estilo.
-¡Ya tienes la clave! Ahora conoces cómo debe ir vestido a una boda un invitado hombre, solo hace falta que apliques tu buen gusto y creatividad para que la combinación de elementos se convierta en una verdadera expresión de tu personalidad. ¿Estás listo para elegir tu look?
-También te puede interesar: 
-También te puede interesar: 
-Tendencias en trajes de novio 2023-2024 vistas en la BBFW, ¡descúbrelas!
-12 accesorios para novio, ¿cuáles incluirás en tu look nupcial?
