--- datasets/La fotografía para la boda.txt
@@ -1,143 +1,0 @@
-La fotografía para la boda
-El día de tu boda será increíble pero pasará muy rápido. El reportaje de fotos inmortalizará los momentos irrepetibles de un día lleno de emociones. Les ayudamos a encontrar el mejor profesional para tener el álbum de boda soñado.
-
-Todos                    
-
-
-El video                    
-
-
-La música                    
-
-
-La decoración                    
-
-
-La papelería                    
-
-
-Los recuerdos                    
-
-
-El transporte                    
-
-
-Tendencias en fotografía de boda: 6 propuestas que no pueden faltar en su gran día            
-
-Estas son las fotos imprescindibles con tu mamá el día de la boda            
-
-Los besos más bonitos de las parejas el día de su boda: ¡vivan los novios!            
-
-¿Que es 'trash the dress'? 15 ideas con fotos originales y todo lo que necesitan saber            
-
-15 sesiones de fotos para boda: ¡descúbranlas todas!            
-
-8 trucos para sacar una foto bonita al anillo de compromiso            
-
-Las fotos con papá que no les pueden faltar en la boda            
-
-Todas las fotos de su boda en un solo álbum: ¡Wedshoots!            
-
-Las poses de parejas famosas para imitar            
-
-Cómo elegir el estilo fotográfico de la boda: ¡6 datos clave para acertar!            
-
-Sesión trash the dress en la playa: las 10 fotos que no pueden faltar            
-
-Fotos con los invitados: ¿retratos de familia mirando a cámara sí o no?            
-
-Best Real Wedding 2021: ¡conozcan el reportaje de boda ganador!            
-
-5 ideas para las fotografías de tu vestido de novia, ¡inspírate!            
-
-75 fotos divertidas con las damas y los 'best men' ¡derechas al álbum!            
-
-12 ideas para un 'save the date' único y original            
-
-Un estilo fotográfico de boda para cada tipo de pareja            
-
-Preparativos de boda                        
-
-Organización de la boda                                    
-
-Actualidad en bodas y expos de boda                                    
-
-Tradiciones y costumbres de boda                                    
-
-Requisitos para matrimonio                                    
-
-Ceremonia de boda                        
-
-Boda religiosa                                    
-
-Boda civil                                    
-
-Lugares para casarse                                    
-
-Protocolo de la boda                                    
-
-Banquete de bodas                        
-
-Lugares para celebrar el banquete                                    
-
-Todo lo que necesitas saber sobre el banquete de boda                                    
-
-Proveedores para tu boda                        
-
-El video de la boda                                    
-
-La música para la boda                                    
-
-La decoración para la boda                                    
-
-Papelería de boda                                    
-
-Recuerdos de boda                                    
-
-El transporte de la boda                                    
-
-Moda en bodas                        
-
-La novia                                    
-
-El novio                                    
-
-Los invitados de la boda                                    
-
-Belleza y salud                        
-
-Belleza para novias                                    
-
-Belleza para novios                                    
-
-Tratamientos                                    
-
-Luna de miel                        
-
-Organización de la luna de miel                                    
-
-Destinos para la luna de miel                                    
-
-Después de la boda                        
-
-Consejos para la vida en pareja                                    
-
-Hazlo tú mism@                        
-
-Historias de boda                        
-
-Agenda de Tareas                        
-
-Mis Invitados                        
-
-Organizador de Mesas                        
-
-Presupuestador                        
-
-Web de Boda                        
-
-
-
-
-
-
