--- datasets/La importancia de los canapés.txt
@@ -1,14 +1,0 @@
-Organización de la boda
-La importancia de los canapés 
-Los aperitivos son parte importante de la boda, recuerden que estos pequeños detalles son los que harán de su evento algo memorable.
-
-Bodas.com.mx            
-
-
-
-Muchas veces tomamos los aperitivos por sentado o simplemente los ignoramos, o colocamos algunas bandejas con queso en las mesas, sin embargo los aperitivos son necesarios para muchas cosas. Por ejemplo, si su recepción es en un lugar diferente al de la ceremonia religiosa, mientras todo se está alistando los meseros empiezan a crear distracción y ambiente para que no empiecen a estresarse con el tiempo. También sirven para empezar cualquier plática entre invitados y calmar esos estómagos que rugen con fuerza.Para tener un éxito garantizado con sus aperitivos, les recomiendo lo siguiente.
-Menos es más.
-del tamaño de un bocado
-Combínelos adecuadamente
-algo dulce
-suficientes meseros
