--- datasets/La luna de miel como regalo de boda.txt
@@ -1,29 +1,0 @@
-Organización de la luna de miel
-La luna de miel como regalo de boda 
-Si son de las parejas que nos les va hacer mesa de regalos con todo tipo de cosas para el hogar y prefieren tomar una merecida luna de miel. Aquí está todo lo que tienen que saber para que sus invitados les regalen la luna de miel como regalo de boda
-
-
-
- 
-Cada día son más las parejas que optan por una luna de miel como regalo de bodas, es por eso que aquí les daremos todos los tips para poder organizarlo.
-luna de miel
-Hagan
-participe a sus invitados
- 
-
- 
-dividan los costos
-invitados
- 
-
-
- 
- las mejores ofertas
-Siempre es bueno ser agradecido, así que no se olviden de sus invitados, ya sea como enviarles alguna postal o hacerles un recuerdo de agradecimiento.  
- 
-Foto 1: Viajemos México
- Viajemos México
-Foto 2: Sevilla Sol Viajes
-Sevilla Sol Viajes
-Foto 3: Excel Tours
-Excel Tours
