--- datasets/¿Serías mi dama?.txt
@@ -1,16 +1,0 @@
-Organización de la boda
-¿Serías mi dama?
-Los pequeños detalles son lo que hacen una boda exitosa y esto incluye desde los preparativos. Aquí te propongo una manera diferente de pedirle a tus amigas que sean tus damas.
-
-Bodas.com.mx            
-
- 
-Las chicas somos más sensibles y cuando tenemos nuestro grupo de amigas hacemos todo lo posible por sentirnos queridas las unas de las otras. Al momento de escoger a tus damas de honor seguro pasas por algunos momentos algo emocionales ya que estas escogiendo a las mujeres más cercanas en tu vida y ellas te apoyarán de muchísimas y diferentes maneras al día de tu boda por eso te propongo que se los pidas de una manera especial.
-Invita a tus amigas a cenar
-Piensa en algún postre
-Ponte de acuerdo con los meseros
-también puedes darles algún detalle
-Asegúrate de no invitar a alguien que no vaya a formar parte de tus damas
-no pueden participar como tus damas
-Foto: Grimaldo FP
-Grimaldo FP
