--- datasets/¿Y si no quiero lanzar el ramo de novia? 10 alternativas muy divertidas.txt
@@ -1,57 +1,0 @@
-Tradiciones y costumbres de boda
-¿Y si no quiero lanzar el ramo de novia? 10 alternativas muy divertidas
-No siempre es del gusto de todas las novias seguir ciertas tradiciones nupciales al pie de la letra. Si eres una de las que se cuestiona por qué lanzar tu ramo de novia, tenemos para ti 10 opciones que te ayudarán a decidir lo mejor para tu boda.
-
-Érika Flores            
-
-
-
-Si ya planeas que tu boda sea muy especial y vanguardista, con un atrevido vestido de novia y un festejo en el que no se lance el ramo de novia, tenemos para ti 10 alternativas que te encantarán. Ya sea porque quieres quedarte con el  natural que tanto te costó elegir o porque te dan alergia las tradiciones, anímate a desafiar la norma.
-ramo de novia
-1. Entrégalo a quien tú quieras
-2. Divídelo entre tus solteras especiales
-3. ¡A cortar listones!
-4. Comparte tu suerte con la siguiente pareja
-5. Regala ramos similares
-6. Ponlo de piñata
-7. Quema la papa…
-8. Descubran quién es la más rápida
-9. A la suerte… o la matemática
-10. Inténtalo como Houdini
-Con el lanzamiento del ramo se busca que la suerte de la novia llegue a manos de una soltera de la que se espera que pronto encuentre el amor o, si ya lo tiene, que lo que llegue sea ese anillo de compromiso a su dedo. Si tú quieres decidir a quién pasas esa fortuna, no dejes al azar que el ramo llegue a las manos correctas.
-que el ramo llegue a las manos correctas
-Al observar tu lista de invitadas te será fácil ubicar a quién deseas transmitir tu buena suerte. Posiblemente sea una amiga muy querida a quien quieras ver tan feliz como lo eres en ese momento. Pero asegúrate de que lo vaya a recibir con ilusión, que el matrimonio esté entre sus planes de futuro y que no vaya a sentirse presionada cuando le pases la estafeta.
-que no vaya a sentirse presionada
-
-
-Entregar una o dos flores de ese codiciado ramo de novia rojo a cada una de tus amigas solteras especiales es compartirles tu dicha y tus mejores deseos de que a ellas también les vaya bien en el amor. Repartirás tu ramo, pero evitarás ponerlo por los aires.
-una o dos flores
-Ya que el lanzamiento del ramo es uno de los momentos más divertidos y esperados en una boda, no desaproveches la oportunidad de plantear un juego. Prepárate con varios listones del mismo tamaño para que los unas alrededor del ramo; llama a las solteras a tomar uno y, sin mirar quién sostiene cada cabo, ve cortando uno a uno hasta llegar al último. Quien tenga el último se quedará con el ansiado premio.
-sin mirar quién sostiene cada cabo, ve cortando uno a uno
-
-
-Si en la planeación de tu boda ya recibiste el  de alguna amistad o de un familiar y no tardarán mucho en llegar sus invitaciones, obsequia tu ramo a esa pareja tan querida que ya anunció su compromiso. Con tus mejores deseos, incluso puedes aprovechar para dar un breve discurso.
-
-Para que no tengas que dar el tuyo, busca en el catálogo de ramos de novia el que creas que es el indicado. Y después llegará el tiempo de compartir con tu amiga, hermana o prima esa gran sabiduría en bodas que tú estás acumulando.
-Si no deseas deshacerte del , ponte de acuerdo con tu proveedor de arreglos florales para que prepare el número que consideres de ramos similares para que, en lugar de lanzarlo, lo entregues personalmente a quien desees. Puedes recurrir a ramos de novia artificiales para que el tuyo se distinga de los obsequiados. Además de que será un detalle muy especial, verás que todas quedarán muy contentas. ¡Imagina la foto con todas las elegidas!
-ramos de novia artificiales
-
-
-Como si fuera este tradicional juego mexicano, el ramo se ata y las candidatas pasan al centro para que, con los ojos vendados, intenten agarrarlo y quedarse con él. Este juego hará estallar las risas. Eso sí, ¡nada de zarandearlo bruscamente!
-tradicional juego mexicano
-No te niegues a la diversión y convoca a las solteras a jugar a la papa caliente. Acomódalas en forma de círculo y haz pasar el ramo de mano en mano al ritmo de tu canción favorita. Sin ver quién lo tiene, cuando creas que es el momento, haz que pare la música y grita “¡se quemó!” La chica que lo tenga en sus manos será la afortunada.
-al ritmo de tu canción favorita
-
-
-Organiza en grupos de cuatro o más personas a las solteras y luego haz el juego de las sillas por cada equipo. Apóyate en el DJ o en el grupo de música para boda que contraten, a fin de que bailen caminando alrededor de las sillas y quienes se vayan quedando sin asiento, dejen la competencia. El grupo se irá reduciendo hasta que queden dos finalistas “peleando” por una sola silla. Quien sea la última en sentarse se queda con el ramo.
-haz el juego de las sillas
-Este puede ser el mejor momento para repartir esas sandalias o balerinas que les tienes preparadas como recuerdos para boda sencillos. Así, la competencia será más justa y nadie se lastimará un tobillo por culpa de los tacones.
-Escoge una flor con muchos pétalos y acomoda a las solteras en círculo o formando una fila. Cada una quitará un pétalo e irá pasando la flor; quien se quede con el último será la ganadora. Pero vigila que nadie hace trampa y arranca varias.
-una flor con muchos pétalos
-
-
-Como el objetivo de este día especial es conseguir que sea inolvidable y pasarla muy bien, sorprende a las solteras incluidas en tu lista de invitados. Coloca el ramo en una caja protegida por cadenas y un candado. Llegado el momento, entrégales una llave distinta a cada una; quien logre abrirla, tendrá fortuna en el amor. Y si son aficionados de los acertijos, las  y los , ¿por qué no ponen a prueba sus destrezas? ¿Quién apuestas que será la primera en conseguir el ramo?
-lista de invitados
-caja protegida por cadenas y un candado
-aficionados de los acertijos, las  y los 
-Como ves, hay distintas formas de resolver este aspecto sin desaprovechar la oportunidad de divertirse y de hacer partícipes a los solteros. Si lo consideras pertinente, avísales la dinámica que aplicarás desde la entrega de las invitaciones de boda, por si alguna quiere ir entrenándose con tiempo. Y si prefieres que sea sorpresa, seguro que será el tema de conversación en cada mesa cuando se sirva el pastel de boda. ¿Qué alternativa elegirás tú?
