--- embeddings/documents_hash.txt
@@ -1,1 +1,0 @@
-7687eb4ec9746bc607d927821c7a21371718ff609f4c11d6d941aff10b6a9e23
